feat: add note type basic and reversed card
This commit adds a boolean parameter `reversed` to the configuration.
If the note type is not cloze and `reversed` is set to `true`, the note type will be set to "Basic (and reversed card)" instead of "Basic".
As this is a standard note type, it should be available on all Anki installations. As the fields are identical to the Basic type, no unexpected problems should arise.